Refactor rewriters into separate files
